+++ WebApplication1/WebApplication1/DTOs/CategoryDTOs.cs
@@ -4,9 +4,9 @@ namespace IncomeExpenseManagementApp.DTOs
 {
     public class CategoryDTO
     {
-        public byte Id { get; set; }
+        public short Id { get; set; }
         public string Name { get; set; } = string.Empty;
-        public byte TransactionTypeId { get; set; }
+        public short TransactionTypeId { get; set; }
         public string? TransactionTypeName { get; set; }
     }
 
@@ -17,6 +17,6 @@ namespace IncomeExpenseManagementApp.DTOs
         public string Name { get; set; } = string.Empty;
 
         [Required]
-        public byte TransactionTypeId { get; set; }
+        public short TransactionTypeId { get; set; }
     }
 }
