+++ WebApplication1/WebApplication1/DTOs/TransactionDTOs.cs
@@ -7,9 +7,9 @@ namespace IncomeExpenseManagementApp.DTOs
         public long Id { get; set; }
         public string Name { get; set; } = string.Empty;
         public string? Description { get; set; }
-        public byte CategoryId { get; set; }
+        public short CategoryId { get; set; }
         public string? CategoryName { get; set; }
-        public byte TransactionTypeId { get; set; }
+        public short TransactionTypeId { get; set; }
         public string? TransactionTypeName { get; set; }
         public decimal Amount { get; set; }
         public DateTime TransactionDate { get; set; }
@@ -25,7 +25,7 @@ namespace IncomeExpenseManagementApp.DTOs
         public string? Description { get; set; }
 
         [Required]
-        public byte CategoryId { get; set; }
+        public short CategoryId { get; set; }
 
         [Required]
         [Range(0.01, 999999999.99)]
@@ -42,7 +42,7 @@ namespace IncomeExpenseManagementApp.DTOs
         [StringLength(500)]
         public string? Description { get; set; }
 
-        public byte? CategoryId { get; set; }
+        public short? CategoryId { get; set; }
 
         [Range(0.01, 999999999.99)]
         public decimal? Amount { get; set; }
