+++ WebApplication1/WebApplication1/Repositories/IEntityRepositories.cs
@@ -9,15 +9,15 @@ namespace IncomeExpenseManagementApp.Repositories
 
     public interface ICategoryRepository : IRepository<Category>
     {
-        Task<IEnumerable<Category>> GetByTransactionTypeAsync(byte transactionTypeId);
-        Task<Category?> GetWithTransactionTypeAsync(byte id);
+        Task<IEnumerable<Category>> GetByTransactionTypeAsync(short transactionTypeId);
+        Task<Category?> GetWithTransactionTypeAsync(short id);
     }
 
     public interface ITransactionRepository : IRepository<Transaction>
     {
         Task<IEnumerable<Transaction>> GetWithCategoryAsync();
         Task<IEnumerable<Transaction>> GetByDateRangeAsync(DateTime startDate, DateTime endDate);
-        Task<IEnumerable<Transaction>> GetByCategoryAsync(byte categoryId);
-        Task<Transaction?> GetWithCategoryByCategoryAsync(byte categoryId);
+        Task<IEnumerable<Transaction>> GetByCategoryAsync(short categoryId);
+        Task<Transaction?> GetWithCategoryByCategoryAsync(short categoryId);
     }
 }
